LayoutTests/webaudio/audiopannernode-basic-expected.txt

Basic test for AudioPannerNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS AudioPannerNode has one input.
PASS AudioPannerNode has one output.
PASS panningModel default value is HRTF.
PASS refDistance default value is 1.
PASS maxDistance default value is 10000.
PASS rolloffFactor default value is 1.
PASS coneInnerAngle default value is 360.
PASS coneOuterAngle default value is 360.
PASS coneOuterGain default value is 0.
PASS panningModel set to EQUALPOWER model and read correctly.
PASS panningModel set to HRTF model and read correctly.
PASS Setting panningModel to SOUNDFIELD correctly throws exception because it is not implemented.
PASS Illegal panningModel correctly throws exception.
PASS distanceModel set to LINEAR_DISTANCE and read correctly.
PASS distanceModel set to INVERSE_DISTANCE and read correctly.
PASS distanceModel set to EXPONENTIAL_DISTANCE and read correctly.
PASS Illegal distanceModel correctly throws exception.
PASS successfullyParsed is true

TEST COMPLETE
